Refactor method names containg Person to Contact
Current state has method names that use Person in the naming.

They do not reflect the changing from Person to the more general Contact
interface.

In order to be more general so that the later ServiceProvider class can
be added, this commit is aimed to fix the overall naming of the methods.
Similar commit messages should be expected doing this task.