Edit DeleteCommandTest/execute_validIndexFilteredList_success()
Previously, this tests for whether the screen updates to a blank screen
after a contact that is filtered is deleted.

However, this is not applicable to Heartsquare anymore as heartsquare
deletes contacts based on unique client/serviceprovider ID.

This commit changes the behaviour such that it tests for the
client/serviceprovider list being shown according to the delete command
called, instead of showing the updated filtered list, which might not be
relevant to the user. E.g:
1. client list
2. serviceprovider#3 delete
The filtered list shown should be serviceprovider list